Remove social interactor and move logic to authenticator implementation
