+++ auth/src/main/java/com/vimeo/networking2/SocialAuthParams.kt
@@ -1,4 +1,4 @@
-package com.vimeo.networking2.internal.interactors
+package com.vimeo.networking2
 
 import com.vimeo.networking2.internal.AuthParam
