+++ auth/src/main/java/com/vimeo/networking2/internal/AuthenticatorImpl.kt
@@ -1,11 +1,6 @@
 package com.vimeo.networking2.internal
 
-import com.vimeo.networking2.AuthCallback
-import com.vimeo.networking2.Authenticator
-import com.vimeo.networking2.GrantType
-import com.vimeo.networking2.VimeoRequest
-import com.vimeo.networking2.internal.interactors.SocialAuthInteractor
-import com.vimeo.networking2.internal.interactors.SocialAuthParams
+import com.vimeo.networking2.*
 
 /**
  * Authentication with email, google, facebook or pincode.
@@ -29,26 +24,35 @@ internal class AuthenticatorImpl(
         return call.enqueueAuthRequest(authCallback)
     }
 
-    /**
-     * Authenticate using Google.
-     */
     override fun google(params: SocialAuthParams, authCallback: AuthCallback): VimeoRequest =
         socialAuthenticate(params, authCallback)
 
-    /**
-     * Authenticate using Facebook.
-     */
     override fun facebook(params: SocialAuthParams, authCallback: AuthCallback): VimeoRequest =
         socialAuthenticate(params, authCallback)
 
     /**
-     * Google and Facebook authentication have the same authentication flows. Both of them require
-     * the client to supply a token and email. For re-usability, both authentication flows use
-     * a single interactor.
+     * Validates the params given the client in [SocialAuthParams]. If they were
+     * invalid values provided to them, it will inform the client of the error. Otherwise,
+     * an API request is made and result is returned to the user.
+     *
+     * @param params        Google or Facebook authentication params such as token, email, etc...
+     * @param authCallback  Callback to inform you of the result of the API request.
      */
     private fun socialAuthenticate(params: SocialAuthParams, authCallback: AuthCallback): VimeoRequest {
-        val interactor = SocialAuthInteractor(authService, authHeaders)
-        return interactor.authenticate(params, authCallback)
+        val invalidAuthParams = params.validate()
+
+        return if (invalidAuthParams.isNotEmpty()) {
+            val apiErrorResponse = createApiErrorForInvalidParams(
+                "Google authentication error.",
+                invalidAuthParams
+            )
+            authCallback.onApiError(apiErrorResponse)
+            NoOpVimeoRequest()
+
+        } else {
+            val call = authService.join(authHeaders, params.toMap())
+            call.enqueueAuthRequest(authCallback)
+        }
     }
 
 }
